PAIN AND SUFFERING

+++ src/clusterfock/cc/rhs/l_QCCD.py
@@ -6,7 +6,7 @@ def lambda_amplitudes_qccd(t2, l2, u, f, o, v):
     dtype = u.dtype
     zeros = lambda shape: np.zeros(shape, dtype=dtype)
 
-    r2 = zeros((M, M, N, N), dtype=u.dtype)
+    r2 = zeros((M, M, N, N))
 
     r2 += np.einsum("aeil,cdkj,cgkl,debg->abij", l2, l2, t2, u[v, v, v, v], optimize=True)
