+++ src/clusterfock/cc/rhs/t_QCCD.py
@@ -6,7 +6,7 @@ def amplitudes_qccd(t2, l2, u, f, v, o):
     dtype = u.dtype
     zeros = lambda shape: np.zeros(shape, dtype=dtype)
 
-    r2 = zeros((M, M, N, N), dtype=u.dtype)
+    r2 = zeros((M, M, N, N))
 
     r2 += np.einsum("cdkl,agil,cekj,bdeg->abij", l2, t2, t2, u[v, v, v, v], optimize=True)
